refactor: extract publish hook
